# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProcureFlow OSS",
    "description": "Open-source AI-assisted RFQ comparison and procurement decision engine.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/quotations": {
      "post": {
        "tags": [
          "quotations"
        ],
        "summary": "Create a new supplier quotation container",
        "operationId": "create_quotation_api_v1_quotations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierQuotationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierQuotationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "quotations"
        ],
        "summary": "List supplier quotations with optional filtering",
        "operationId": "list_quotations_api_v1_quotations_get",
        "parameters": [
          {
            "name": "rfq_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by RFQ ID",
              "title": "Rfq Id"
            },
            "description": "Filter by RFQ ID"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QuotationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierQuotationRead"
                  },
                  "title": "Response List Quotations Api V1 Quotations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotations/upload": {
      "post": {
        "tags": [
          "quotations"
        ],
        "summary": "Upload New Quotation",
        "operationId": "upload_new_quotation_api_v1_quotations_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_new_quotation_api_v1_quotations_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierQuotationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotations/{quotation_id}": {
      "get": {
        "tags": [
          "quotations"
        ],
        "summary": "Get quotation details and attached documents",
        "operationId": "get_quotation_api_v1_quotations__quotation_id__get",
        "parameters": [
          {
            "name": "quotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quotation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierQuotationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotations/{quotation_id}/documents": {
      "post": {
        "tags": [
          "quotations"
        ],
        "summary": "Upload and immutably attach a quotation document (PDF, XLSX, CSV)",
        "operationId": "upload_quotation_document_api_v1_quotations__quotation_id__documents_post",
        "parameters": [
          {
            "name": "quotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quotation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_quotation_document_api_v1_quotations__quotation_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotations/{quotation_id}/documents/{document_id}/download": {
      "get": {
        "tags": [
          "quotations"
        ],
        "summary": "Download the original immutable supplier quotation document",
        "operationId": "download_quotation_document_api_v1_quotations__quotation_id__documents__document_id__download_get",
        "parameters": [
          {
            "name": "quotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quotation Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotations/{quotation_id}/extract": {
      "post": {
        "tags": [
          "quotations"
        ],
        "summary": "Trigger or retry asynchronous document extraction",
        "operationId": "trigger_extraction_api_v1_quotations__quotation_id__extract_post",
        "parameters": [
          {
            "name": "quotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quotation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotations/{quotation_id}/extractions/latest": {
      "get": {
        "tags": [
          "quotations"
        ],
        "summary": "Get current authoritative extraction result with line items, coordinates, and warnings",
        "operationId": "get_latest_extraction_api_v1_quotations__quotation_id__extractions_latest_get",
        "parameters": [
          {
            "name": "quotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quotation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractedQuotationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotations/{quotation_id}/line-items": {
      "post": {
        "tags": [
          "quotations"
        ],
        "summary": "Add a missed line item to an extraction",
        "operationId": "create_line_item_api_v1_quotations__quotation_id__line_items_post",
        "parameters": [
          {
            "name": "quotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quotation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractedLineItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractedLineItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotations/{quotation_id}/line-items/{line_item_id}": {
      "patch": {
        "tags": [
          "quotations"
        ],
        "summary": "Human-in-the-loop correction of an extracted line item",
        "operationId": "correct_line_item_api_v1_quotations__quotation_id__line_items__line_item_id__patch",
        "parameters": [
          {
            "name": "quotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quotation Id"
            }
          },
          {
            "name": "line_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Line Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractedLineItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractedLineItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "quotations"
        ],
        "summary": "Soft-delete/exclude an extracted line item (never hard-deletes parser items)",
        "operationId": "soft_delete_line_item_api_v1_quotations__quotation_id__line_items__line_item_id__delete",
        "parameters": [
          {
            "name": "quotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quotation Id"
            }
          },
          {
            "name": "line_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Line Item Id"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for exclusion",
              "title": "Reason"
            },
            "description": "Reason for exclusion"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractedLineItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotations/{quotation_id}/line-items/{line_item_id}/restore": {
      "post": {
        "tags": [
          "quotations"
        ],
        "summary": "Restore a previously excluded line item",
        "operationId": "restore_line_item_api_v1_quotations__quotation_id__line_items__line_item_id__restore_post",
        "parameters": [
          {
            "name": "quotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quotation Id"
            }
          },
          {
            "name": "line_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Line Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractedLineItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotations/{quotation_id}/fields/{field_id}": {
      "patch": {
        "tags": [
          "quotations"
        ],
        "summary": "Human-in-the-loop correction of a quotation header field",
        "operationId": "correct_quotation_field_api_v1_quotations__quotation_id__fields__field_id__patch",
        "parameters": [
          {
            "name": "quotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quotation Id"
            }
          },
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractedFieldUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractedFieldRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuotationStatus": {
        "type": "string",
        "enum": [
          "uploaded",
          "queued",
          "extracting",
          "needs_review",
          "approved",
          "rejected",
          "failed"
        ],
        "title": "QuotationStatus"
      },
...
      "SupplierQuotationCreate": {
        "properties": {
          "supplier_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Supplier Name"
          },
          "supplier_reference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Reference"
          },
          "rfq_id": {
            "type": "string",
            "title": "Rfq Id"
          }
        },
        "type": "object",
        "required": [
          "supplier_name",
          "rfq_id"
        ],
        "title": "SupplierQuotationCreate"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}